Updated OptionManager
Options now work much differently, with a class, or instance, required
to "register" itself with the OptionManager for it's options to appear
on the screen

The ChunkTester now has options for FPSCap and VSync, mostly used during
testing to make sure everything worked fine, which it did!

"Voxel Creation Description.odg" This is a graphical explanation of how
the chunk mesh is built, explains it far better that I ever could in
words.